refactor(kolme): extract fork finality resolver module ownership (#1944)
