+++ crates/kamn-core/src/kolme_runtime_commit.rs
@@ -110,6 +110,7 @@ use std::process::{Command, Stdio};
 use std::thread;
 use std::time::{Duration, Instant};
 
+mod fork_finality_resolver;
 mod in_memory_client;
 mod notifications_consumer;
 mod notifications_websocket;
@@ -502,6 +503,7 @@ pub enum KolmeRuntimeCommitOutcome {
 /// Runtime lifecycle state projected from commit receipt outcomes.
 pub type RuntimeCommitLifecycleState = KamnKolmeRuntimeCommitLifecycleState;
 
+pub use fork_finality_resolver::KolmeRuntimeCommitForkFinalityResolver;
 pub use in_memory_client::InMemoryKolmeRuntimeCommitClient;
 pub use notifications_consumer::KolmeRuntimeCommitNotificationsConsumer;
 pub use notifications_websocket::{
@@ -1367,83 +1369,6 @@ impl<T: KolmeRuntimeCommitBlockFallbackTransport> KolmeRuntimeCommitBlockFallbac
                 latest_height,
             ),
         })
-    }
-}
-
-/// Fork-profile finality resolver that composes notifications and block fallback lookups.
-pub struct KolmeRuntimeCommitForkFinalityResolver<C, T>
-where
-    C: KolmeRuntimeCommitNotificationsConnector,
-    T: KolmeRuntimeCommitBlockFallbackTransport,
-{
-    notifications_consumer: KolmeRuntimeCommitNotificationsConsumer<C>,
-    block_fallback_reconciler: KolmeRuntimeCommitBlockFallbackReconciler<T>,
-}
-
-impl<C, T> KolmeRuntimeCommitForkFinalityResolver<C, T>
-where
-    C: KolmeRuntimeCommitNotificationsConnector,
-    T: KolmeRuntimeCommitBlockFallbackTransport,
-{
-    /// Builds a fork finality resolver from notifications and block fallback components.
-    pub fn new(
-        notifications_consumer: KolmeRuntimeCommitNotificationsConsumer<C>,
-        block_fallback_reconciler: KolmeRuntimeCommitBlockFallbackReconciler<T>,
-    ) -> Self {
-        Self {
-            notifications_consumer,
-            block_fallback_reconciler,
-        }
-    }
-
-    /// Resolves finality for one commit id using notifications first, then bounded block fallback.
-    pub fn resolve_commit_finality(
-        &mut self,
-        commit_id: &str,
-        from_height: u64,
-        latest_height: u64,
-    ) -> Result<KolmeRuntimeCommitProviderReceipt, KolmeRuntimeCommitProviderError> {
-        let expected_txhash = txhash_from_kolme_commit_id(commit_id).map_err(|error| {
-            KolmeRuntimeCommitProviderError::MalformedResponse {
-                reason: error.to_string(),
-            }
-        })?;
-
-        match self.notifications_consumer.next_notification_event() {
-            Ok(event) => match event {
-                KolmeRuntimeCommitNotificationEvent::LatestBlock { height } => {
-                    let upper_bound =
-                        resolve_kolme_lookup_upper_bound(from_height, latest_height, height);
-                    self.block_fallback_reconciler.reconcile_txhash(
-                        expected_txhash.as_str(),
-                        from_height,
-                        upper_bound,
-                    )
-                }
-                _ => {
-                    let receipt = event
-                        .to_provider_receipt(self.notifications_consumer.provider())
-                        .ok_or_else(|| KolmeRuntimeCommitProviderError::MalformedResponse {
-                            reason: "notification event did not carry receipt data".to_owned(),
-                        })?;
-                    require_kolme_commit_id_matches_expected_txhash_contract(
-                        receipt.commit_id.as_str(),
-                        expected_txhash.as_str(),
-                    )
-                    .map_err(|error| {
-                        KolmeRuntimeCommitProviderError::MalformedResponse {
-                            reason: error.to_string(),
-                        }
-                    })?;
-                    Ok(receipt)
-                }
-            },
-            Err(KolmeRuntimeCommitProviderError::Unavailable { .. })
-            | Err(KolmeRuntimeCommitProviderError::Timeout) => self
-                .block_fallback_reconciler
-                .reconcile_txhash(expected_txhash.as_str(), from_height, latest_height),
-            Err(error @ KolmeRuntimeCommitProviderError::MalformedResponse { .. }) => Err(error),
-        }
     }
 }
